Update comments
The function canTransform is used to judge whether can carry out this transformation, this error should be "can not find transform frame". The lookupTransform is used to find the latest, valid and avaliable transformation, if it returns empty, you can output the tips that "can't find the transform". I think you can find some similar cases in ROS system.