use supervisord for services in dev environment
This reverts https://github.com/ApolloAuto/apollo/commit/67fad1f572
commit. Without supervisord, once some module dies/quits then ROS
master still thinks the node is present, and any further interaction
with it will fail and ROS commands will behave in very strange way.
This is what makes bridge to die when ros nodes are not correctly
running. With this patch, the ROS nodes are restarted as soon as they
die and ROS clients can successfully connect to them.

For proper fix the ROS itself needs to be fixed (more specifically -
Apollo modifications to ROS). It should detect that process for ROS node
is terminated so ROS node should not be announced.